chore: 계층 분리 - 도메인 계층에 의존 관계를 가지는 경우 infra 계층으로 분리
