+++ .gitignore
@@ -1,6 +1,7 @@
 HELP.md
 .gradle
 build/
+/src/main/resources/data
 !gradle/wrapper/gradle-wrapper.jar
 !**/src/main/**/build/
 !**/src/test/**/build/
