+++ src/main/java/com/bob/domain/post/entity/Post.java
@@ -1,5 +1,6 @@
 package com.bob.domain.post.entity;
 
+import com.bob.domain.trade.entity.Trade;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
@@ -10,7 +11,10 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
+import java.util.ArrayList;
+import java.util.List;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -71,4 +75,7 @@ public class Post extends BaseTime {
   @Column(nullable = false)
   @Builder.Default
   private Integer scrapCount = 0;
+
+  @OneToMany(mappedBy = "post")
+  private final List<Trade> trades = new ArrayList<>();
 }
