--- src/main/java/com/bob/global/auth/filter/request/LoginRequest.java
@@ -1,7 +1,0 @@
-package com.bob.global.auth.filter.request;
-
-public record LoginRequest(
-    String email,
-    String password
-) {
-}
