--- src/main/java/com/bob/global/auth/service/MemberDetailsService.java
@@ -1,26 +1,0 @@
-package com.bob.global.auth.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-import com.bob.domain.member.entity.Member;
-import com.bob.domain.member.repository.MemberRepository;
-import com.bob.global.auth.response.MemberDetails;
-
-@Service
-@RequiredArgsConstructor
-public class MemberDetailsService implements UserDetailsService {
-
-  private final MemberRepository memberRepository;
-
-  @Override
-  public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-    Member member = memberRepository.findByEmail(username)
-        .orElseThrow(() -> new UsernameNotFoundException(username));
-
-    return new MemberDetails(member.getId(), member.getPassword(), member.getEmail());
-  }
-
-}
