--- src/main/java/com/bob/global/config/RedisConfig.java
@@ -1,36 +1,0 @@
-package com.bob.global.config;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.connection.RedisConnectionFactory;
-import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
-import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
-import org.springframework.data.redis.core.StringRedisTemplate;
-import org.springframework.data.redis.serializer.StringRedisSerializer;
-
-@Configuration
-public class RedisConfig {
-
-  @Value("${spring.data.redis.host}")
-  private String redisHost;
-
-  @Value("${spring.data.redis.port}")
-  private int redisPort;
-
-  @Bean
-  public RedisConnectionFactory redisConnectionFactory() {
-    RedisStandaloneConfiguration redisConfiguration = new RedisStandaloneConfiguration();
-    redisConfiguration.setHostName(redisHost);
-    redisConfiguration.setPort(redisPort);
-    return new LettuceConnectionFactory(redisConfiguration);
-  }
-
-  @Bean
-  public StringRedisTemplate stringRedisTemplate(RedisConnectionFactory connectionFactory) {
-    StringRedisTemplate template = new StringRedisTemplate(connectionFactory);
-    template.setKeySerializer(new StringRedisSerializer());
-    template.setValueSerializer(new StringRedisSerializer());
-    return template;
-  }
-}
