+++ src/main/java/com/bob/infra/auth/filter/request/LoginRequest.java
@@ -1,0 +1,7 @@
+package com.bob.infra.auth.filter.request;
+
+public record LoginRequest(
+    String email,
+    String password
+) {
+}
