+++ src/main/java/com/bob/infra/auth/service/MemberDetailsService.java
@@ -1,0 +1,26 @@
+package com.bob.infra.auth.service;
+
+import com.bob.infra.auth.response.MemberDetails;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+import com.bob.domain.member.entity.Member;
+import com.bob.domain.member.repository.MemberRepository;
+
+@Service
+@RequiredArgsConstructor
+public class MemberDetailsService implements UserDetailsService {
+
+  private final MemberRepository memberRepository;
+
+  @Override
+  public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+    Member member = memberRepository.findByEmail(username)
+        .orElseThrow(() -> new UsernameNotFoundException(username));
+
+    return new MemberDetails(member.getId(), member.getPassword(), member.getEmail());
+  }
+
+}
