+++ src/main/java/com/bob/web/member/controller/MemberController.java
@@ -1,5 +1,9 @@
 package com.bob.web.member.controller;
 
+import com.bob.domain.member.service.MemberService;
+import com.bob.web.common.CommonResponse;
+import com.bob.web.common.symbol.ResponseSymbol;
+import com.bob.web.member.request.SignupRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -7,10 +11,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import com.bob.domain.member.service.MemberService;
-import com.bob.web.common.CommonResponse;
-import com.bob.web.common.symbol.ResponseSymbol;
-import com.bob.web.member.request.SignupRequest;
 
 @RequiredArgsConstructor
 @RequestMapping("/members")
@@ -19,9 +19,9 @@ public class MemberController {
 
   private final MemberService memberService;
 
-  @PostMapping("/signup")
+  @PostMapping
   @ResponseStatus(HttpStatus.CREATED)
-  public CommonResponse<ResponseSymbol> singup(@RequestBody SignupRequest request) {
+  public CommonResponse<ResponseSymbol> handleSignup(@RequestBody SignupRequest request) {
     memberService.signupProcess(request.toCommand());
     return new CommonResponse<>(true, ResponseSymbol.CREATED);
   }
