--- src/main/resources/data/category_data.sql
@@ -1,60 +1,0 @@
-INSERT INTO categories (id, name, parent_id) VALUES (1, '경제/경영', NULL);
-INSERT INTO categories (id, name, parent_id) VALUES (2, '과학/IT', NULL);
-INSERT INTO categories (id, name, parent_id) VALUES (3, '만화', NULL);
-INSERT INTO categories (id, name, parent_id) VALUES (4, '문학', NULL);
-INSERT INTO categories (id, name, parent_id) VALUES (5, '어린이/청소년', NULL);
-INSERT INTO categories (id, name, parent_id) VALUES (6, '예술/문화', NULL);
-INSERT INTO categories (id, name, parent_id) VALUES (7, '인문/사회', NULL);
-INSERT INTO categories (id, name, parent_id) VALUES (8, '자기계발', NULL);
-INSERT INTO categories (id, name, parent_id) VALUES (9, '취미/실용', NULL);
-INSERT INTO categories (id, name, parent_id) VALUES (10, '학습/수험', NULL);
-INSERT INTO categories (id, name, parent_id) VALUES (11, '기타', NULL);
-INSERT INTO categories (id, name, parent_id) VALUES (12, '경영일반', 1);
-INSERT INTO categories (id, name, parent_id) VALUES (13, '마케팅/세일즈', 1);
-INSERT INTO categories (id, name, parent_id) VALUES (14, '재테크', 1);
-INSERT INTO categories (id, name, parent_id) VALUES (15, '창업', 1);
-INSERT INTO categories (id, name, parent_id) VALUES (16, '회계/재무', 1);
-INSERT INTO categories (id, name, parent_id) VALUES (17, '공학일반', 2);
-INSERT INTO categories (id, name, parent_id) VALUES (18, '수학', 2);
-INSERT INTO categories (id, name, parent_id) VALUES (19, '자연과학', 2);
-INSERT INTO categories (id, name, parent_id) VALUES (20, '컴퓨터공학', 2);
-INSERT INTO categories (id, name, parent_id) VALUES (21, '프로그래밍', 2);
-INSERT INTO categories (id, name, parent_id) VALUES (22, '그래픽노블', 3);
-INSERT INTO categories (id, name, parent_id) VALUES (23, '웹툰', 3);
-INSERT INTO categories (id, name, parent_id) VALUES (24, '일본만화', 3);
-INSERT INTO categories (id, name, parent_id) VALUES (25, '한국만화', 3);
-INSERT INTO categories (id, name, parent_id) VALUES (26, '시/희곡', 4);
-INSERT INTO categories (id, name, parent_id) VALUES (27, '에세이', 4);
-INSERT INTO categories (id, name, parent_id) VALUES (28, '외국소설', 4);
-INSERT INTO categories (id, name, parent_id) VALUES (29, '한국소설', 4);
-INSERT INTO categories (id, name, parent_id) VALUES (30, '동화책', 5);
-INSERT INTO categories (id, name, parent_id) VALUES (31, '유아그림책', 5);
-INSERT INTO categories (id, name, parent_id) VALUES (32, '청소년 문학', 5);
-INSERT INTO categories (id, name, parent_id) VALUES (33, '학습만화', 5);
-INSERT INTO categories (id, name, parent_id) VALUES (34, '미술/디자인', 6);
-INSERT INTO categories (id, name, parent_id) VALUES (35, '사진', 6);
-INSERT INTO categories (id, name, parent_id) VALUES (36, '영화/연극', 6);
-INSERT INTO categories (id, name, parent_id) VALUES (37, '음악/악보', 6);
-INSERT INTO categories (id, name, parent_id) VALUES (38, '사회학', 7);
-INSERT INTO categories (id, name, parent_id) VALUES (39, '심리학', 7);
-INSERT INTO categories (id, name, parent_id) VALUES (40, '역사', 7);
-INSERT INTO categories (id, name, parent_id) VALUES (41, '정치/법', 7);
-INSERT INTO categories (id, name, parent_id) VALUES (42, '철학', 7);
-INSERT INTO categories (id, name, parent_id) VALUES (43, '리더십', 8);
-INSERT INTO categories (id, name, parent_id) VALUES (44, '성공/동기부여', 8);
-INSERT INTO categories (id, name, parent_id) VALUES (45, '습관', 8);
-INSERT INTO categories (id, name, parent_id) VALUES (46, '시간관리', 8);
-INSERT INTO categories (id, name, parent_id) VALUES (47, '인간관계', 8);
-INSERT INTO categories (id, name, parent_id) VALUES (48, '건강', 9);
-INSERT INTO categories (id, name, parent_id) VALUES (49, '여행', 9);
-INSERT INTO categories (id, name, parent_id) VALUES (50, '요리', 9);
-INSERT INTO categories (id, name, parent_id) VALUES (51, '육아', 9);
-INSERT INTO categories (id, name, parent_id) VALUES (52, '인테리어', 9);
-INSERT INTO categories (id, name, parent_id) VALUES (53, '공무원/수험서', 10);
-INSERT INTO categories (id, name, parent_id) VALUES (54, '대학교재', 10);
-INSERT INTO categories (id, name, parent_id) VALUES (55, '문제집', 10);
-INSERT INTO categories (id, name, parent_id) VALUES (56, '자격증', 10);
-INSERT INTO categories (id, name, parent_id) VALUES (57, '초·중·고 참고서', 10);
-INSERT INTO categories (id, name, parent_id) VALUES (58, '기타도서', 11);
-INSERT INTO categories (id, name, parent_id) VALUES (59, '잡지', 11);
-INSERT INTO categories (id, name, parent_id) VALUES (60, '정기간행물', 11);
