+++ src/test/intg/java/com/bob/support/redis/RedisContainerConfig.java
